refactor: align auth with new backend
