+++ src/app/general/layout.tsx
@@ -18,8 +18,7 @@ export default function GeneralLayout({
   useEffect(() => {
     getMe()
       .then((user) => {
-        const roles: string[] = user?.roles ??
-          user?.rol_usuario?.map((r: any) => r?.rol?.nombre) ?? [];
+        const roles: string[] = user?.roles ?? [];
         if (roles.includes('ADMIN')) {
           router.replace('/admin/asignaciones');
         } else if (roles.includes('SUPERVISOR')) {
